+++ CHANGELOG_PENDING.md
@@ -2,3 +2,5 @@
 
 ### Bug Fixes
 
+- Buffer output to the system pager in paged commands.
+  [#327](https://github.com/pulumi/esc/pull/327)
